+++ jock/git.py
@@ -3,25 +3,26 @@ import sys
 
 import click
 
+from jock.constant import ADDRESS, LOCATION, GIT, COMMAND_PATH
 from jock.utils import get_repository_path
 
 
 def git_common(command, selected_repositories, git_args=()):
     for repository_name in selected_repositories:
         repository = selected_repositories[repository_name]
-        repository_path = get_repository_path(repository['location'])
+        repository_path = get_repository_path(repository[LOCATION])
         click.echo('Executing [{}] in [{}]'.format(command, repository_path))
-        subprocess.run(('git', '-C', repository_path, command) + git_args)
+        subprocess.run((GIT, COMMAND_PATH, repository_path, command) + git_args)
 
 
 def git_clone(selected_repositories, git_args=()):
     for repository_name in selected_repositories:
         repository = selected_repositories[repository_name]
-        repository_path = get_repository_path(repository['location'])
+        repository_path = get_repository_path(repository[LOCATION])
         click.echo(
             'Cloning [{}] in [{}]'.format(repository_name, repository_path)
         )
-        subprocess.run(('git', 'clone', repository['address'], repository_path) + git_args)
+        subprocess.run((GIT, 'clone', repository[ADDRESS], repository_path) + git_args)
 
 
 GIT_COMMANDS = {