+++ CHANGELOG.md
@@ -8,6 +8,15 @@ Note, "non-notable" changes may be small patches with no noticeable effect to th
 
 ## Unreleased
 
+## 0.3.0 2021-03-25
+### Added
+- Remote Imports
+    - `imports` can be defined in `~/.jockrc` with a remote git `address`
+    - `-i`/`--import-config` will fetch the remote `.jockrc` file and store the repositories & groups under `data`
+    - Imported repositories & groups can be used as normal
+
+### Changed
+- Improved error messages
 
 ## 0.2.0 2020-12-17
 ### Added